requirements.txt
Pillow
xxhash
beautifulsoup4
validators
pyperclip
opencv-contrib-python-headless
cached-property